avm2:Move abstract_class_allocator to object.rs
